+++ internal/env/notification_channels.go
@@ -13,6 +13,7 @@ type NotificationChannels struct {
 	TelegramChannels map[string]*notifiler.Telegram
 	DiscordChannels  map[string]*notifiler.Discord
 	OpsGenieChannels map[string]*notifiler.OpsGenie
+	SlackChannels    map[string]*notifiler.Slack
 }
 
 func NewNotificationChannels(
@@ -21,11 +22,13 @@ func NewNotificationChannels(
 	httpClient *http.Client,
 	metricsStore *metrics.Store,
 	blockExplorer, source string,
+	env string,
 ) (*NotificationChannels, error) {
 	channels := &NotificationChannels{
 		TelegramChannels: make(map[string]*notifiler.Telegram),
 		DiscordChannels:  make(map[string]*notifiler.Discord),
 		OpsGenieChannels: make(map[string]*notifiler.OpsGenie),
+		SlackChannels:    make(map[string]*notifiler.Slack),
 	}
 
 	for _, tgChannel := range cfg.TelegramChannels {
@@ -58,13 +61,25 @@ func NewNotificationChannels(
 			metricsStore,
 			source,
 			blockExplorer,
+			env,
 		)
 		log.Info(fmt.Sprintf("Initialized %s channel: %s", opsGenieChannel.ID, opsGenieChannel.Description))
+	}
+
+	for _, slackChannel := range cfg.SlackChannels {
+		channels.SlackChannels[slackChannel.ID] = notifiler.NewSlack(
+			slackChannel.WebhookURL,
+			httpClient,
+			metricsStore,
+			source,
+			blockExplorer,
+		)
+		log.Info(fmt.Sprintf("Initialized %s channel: %s", slackChannel.ID, slackChannel.Description))
 	}
 
 	return channels, nil
 }
 
 func (n *NotificationChannels) Count() int {
-	return len(n.TelegramChannels) + len(n.DiscordChannels) + len(n.OpsGenieChannels)
+	return len(n.TelegramChannels) + len(n.DiscordChannels) + len(n.OpsGenieChannels) + len(n.SlackChannels)
 }